(fix): organize sidebar SVG
