+++ __pages__/employees/employee-details.test.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import EmployeeDetails from './../index';
-
+import EmployeeDetails from '../../pages/employees/employee-details';
 
 it('renders correctly', () => {
-  const Employeedetails = renderer.create( <EmployeeDetails /> );
+  const Employeedetails = renderer.create(<EmployeeDetails />);
   expect(Employeedetails).toMatchSnapshot();
 });
