--- pages/employees/SearchEmployee.tsx
@@ -1,64 +1,0 @@
-import type { NextPage } from "next";
-import Head from "next/head";
-import SearchInput from "../../dashboard-assets/search.svg";
-import Filter from "../../dashboard-assets/filter.svg";
-import EmployeeInfo from "../../dashboard-assets/employeeIcon.svg";
-import Image from "next/image";
-
-const SearchEmployee: NextPage = () => {
-  return (
-    <div>
-      <div className="employee-section">
-        <Head>
-          <title>Employee</title>
-          <meta name="description" content="Generated by create next app" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-        <div className="employee-section">
-          <div className="employee-section__employeeSearch">
-            <div className="employee-section__employeeHeader">
-              {/* <p>30 Employee(s)</p> */}
-              <p>0 Employee(s)</p>
-            </div>
-            <div className="employee-section__searchInput">
-              <div className="employee-section__searchIcon">
-                <div>
-                  <input
-                    type="text"
-                    placeholder="Search by name"
-                    className="employee-section__search"
-                  />
-                </div>
-                <div className="employee-section__searchImage">
-                  <Image src={SearchInput} alt="search-image" />
-                </div>
-                <div className="employee-section__searchIcon">
-                  <div>
-                    <input
-                      type="text"
-                      placeholder="Filter"
-                      className="employee-section__filter"
-                    />
-                  </div>
-                  <div className="employee-section__searchImage">
-                    <Image src={Filter} alt="filter-image" />
-                  </div>
-                  <button
-                    type="submit"
-                    className="employee-section__employeeInfo">
-                    <Image src={EmployeeInfo} alt="info-image" />
-                  </button>
-                  {/* <button className="employee-section__employeeInfo">
-                    <Image src={EmployeeInfo} alt="info-image" />
-                  </button> */}
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default SearchEmployee;
