--- pages/employees/employeeTableDetails.tsx
@@ -1,205 +1,0 @@
-import type { NextPage } from "next";
-import Head from "next/head";
-import Image from "next/image";
-// import Inbox from "../../public/employee-inbox.svg";
-import EmployeeInfo from "../../dashboard-assets/employeeIcon.svg"
-
-const employeeTableDetails: NextPage =  () => {
-  return (
-    <div>
-      <div className="employee-section">
-        <Head>
-          <title>EmployeeTableDetails</title>
-          <meta name="description" content="Generated by create next app" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-      </div>
-      <section>
-        <table>
-          <tr className="employee-section__employeeTitle">
-            <span>
-              <input
-                type="checkbox"
-                className="employee-section__employee_Input"
-                value="checkinputOne"
-              />
-              <span className="employee-section__employee-name">
-                {' '}
-                <th>Name</th>
-              </span>
-            </span>
-            <th>Email Address</th>
-            <th>Amount (₦) </th>
-            <th>Payout Method</th>
-            <th>Group</th>
-            <th>Date Added</th>
-          </tr>
-          <tr className="employee-section__employeeData">
-            <span>
-              <input
-              type="checkbox"
-              className="employee-section__employee_Input"
-              value="checkInputTwo"
-              />
-              <td>Esther Howard</td>
-            </span>
-            <td>estherhoward@gmail.com</td>
-            <td>₦ 120,000</td>
-            <td>Bank Account</td>
-            <td>Payroll group 1</td>
-            <td>May 27, 2020 | <span className="employee-section__employee_pay-time">12:38 PM</span> </td>
-            <div className="employee-section__employeeImage ">
-              <Image src={EmployeeInfo} alt="employee-image" />
-              </div>
-          </tr>
-          <tr className="employee-section__employeeData">
-            <span>
-            <input
-                type="checkbox"
-                className="employee-section__employee_Input"
-                value="checkInputThree"
-                />
-              <td>Esther Howard</td>
-            </span>
-            <td>estherhoward@gmail.com</td>
-            <td>₦ 120,000</td>
-            <td>Bank Account</td>
-            <td>Payroll group 1</td>
-            <td>May 27, 2020  <span className="employee-section__employee_pay-time">12:38 PM</span></td>
-            <div className="employee-section__employeeImage ">
-            <Image src={EmployeeInfo} alt="employee-image" />
-            </div>
-            </tr>
-          <tr className="employee-section__employeeData">
-          <span>
-              <input
-              type="checkbox"
-                className="employee-section__employee_Input"
-                value="checkInputFour"
-                />
-              <td>Esther Howard</td>
-              </span>
-            <td>estherhoward@gmail.com</td>
-            <td>₦ 120,000</td>
-            <td>Bank Account</td>
-            <td>Payroll group 1</td>
-            <td>May 27, 2020 | <span className="employee-section__employee_pay-time">12:38 PM</span></td>
-            <div className="employee-section__employeeImage ">
-              <Image src={EmployeeInfo} alt="employee-image" />
-              </div>
-          </tr>
-          <tr className="employee-section__employeeData">
-          <span>
-              <input
-              type="checkbox"
-              className="employee-section__employee_Input"
-                value="checkInputFive"
-                />
-              <td>Esther Howard</td>
-              </span>
-            <td>estherhoward@gmail.com</td>
-            <td>₦ 120,000</td>
-            <td>Bank Account</td>
-            <td>Payroll group 1</td>
-            <td>May 27, 2020 |  <span className="employee-section__employee_pay-time">12:38 PM</span></td>
-            <div className="employee-section__employeeImage ">
-              <Image src={EmployeeInfo} alt="employee-image" />
-            </div>
-          </tr>
-          <tr className="employee-section__employeeData">
-            <span>
-              <input
-                type="checkbox"
-                className="employee-section__employee_Input"
-                value="checkInputSix"
-              />
-              <td>Esther Howard</td>
-            </span>
-            <td>estherhoward@gmail.com</td>
-            <td>₦ 120,000</td>
-            <td>Bank Account</td>
-            <td>Payroll group 1</td>
-            <td>May 27, 2020 |  <span className="employee-section__employee_pay-time">12:38 PM</span></td>
-            <div className="employee-section__employeeImage ">
-              <Image src={EmployeeInfo} alt="employee-image" />
-            </div>
-            </tr>
-
-          <tr className="employee-section__employeeData">
-          <span>
-              <input
-                type="checkbox"
-                className="employee-section__employee_Input"
-                value="checkInputSix"
-              />
-              <td>Esther Howard</td>
-              </span>
-            <td>estherhoward@gmail.com</td>
-            <td>₦ 120,000</td>
-            <td>Bank Account</td>
-            <td>Payroll group 1</td>
-            <td>May 27, 2020 |  <span className="employee-section__employee_pay-time">12:38 PM</span></td>
-            <div className="employee-section__employeeImage ">
-              <Image src={EmployeeInfo} alt="employee-image" />
-            </div>
-          </tr>
-
-          <tr className="employee-section__employeeData">
-            <span>
-            <input
-                type="checkbox"
-                className="employee-section__employee_Input"
-                value="checkInputSix"
-              />
-              <td>Esther Howard</td>
-            </span>
-            <td>estherhoward@gmail.com</td>
-            <td>₦ 120,000</td>
-            <td>Bank Account</td>
-            <td>Payroll group 1</td>
-            <td>May 27, 2020 | <span className="employee-section__employee_pay-time">12:38 PM</span> </td>
-            <div className="employee-section__employeeImage ">
-              <Image src={EmployeeInfo} alt="employee-image" />
-              </div>
-              </tr>
-
-              <tr className="employee-section__employeeData">
-            <span>
-              <input
-              type="checkbox"
-                className="employee-section__employee_Input"
-                value="checkInputSix"
-                />
-              <td>Esther Howard</td>
-              </span>
-            <td>estherhoward@gmail.com</td>
-            <td>₦ 120,000</td>
-            <td>First Bank Nig</td>
-            <td>Payroll group 1</td>
-            <td>May 27, 2020 |  <span className="employee-section__employee_pay-time">12:38 PM</span></td>
-            <div className="employee-section__employeeImage ">
-            <Image src={EmployeeInfo} alt="employee-image" />
-            </div>
-          </tr>
-        </table>
-            {/* <div className="employee-section__employee-inbox">
-              <Image src={Inbox} />
-              <p>No employee yet</p>
-            </div> */}
-        <div className="employee-section__nextPages">
-          <p>
-            Showing <span className="employee">Page 1 of 10</span>
-          </p>
-          <div className="employee-section__prev_next_pages">
-            <p>Prev</p>
-            <p>1</p>
-            <p>2</p>
-            <p>Next</p>
-          </div>
-        </div>
-      </section>
-    </div>
-  );
-};
-
-export default employeeTableDetails;
