+++ src/layouts/dashboard-layout/DashBoardLayout.tsx
@@ -11,6 +11,8 @@ const DashboardLayout: React.FC = ({ children }) => {
 
   const dashboardLink = '/dashboard' || '/';
 
+  console.log('Route name: ', router.pathname);
+
   return (
     <div className="dashboardLayout">
       <nav className="dashboard-navigation">
@@ -29,30 +31,7 @@ const DashboardLayout: React.FC = ({ children }) => {
                   router.pathname === dashboardLink ? 'active' : '',
                 ].join(' ')}
               >
-                <svg
-                  width="20"
-                  height="20"
-                  viewBox="0 0 20 20"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M4.64286 2.5C3.45939 2.5 2.5 3.45939 2.5 4.64286V8.48214C2.5 9.66561 3.45939 10.625 4.64286 10.625H6.78571C7.96918 10.625 8.92857 9.66561 8.92857 8.48214V4.64286C8.92857 3.45939 7.96918 2.5 6.78571 2.5H4.64286Z"
-                    fill="currentColor"
-                  />
-                  <path
-                    d="M4.64286 12.2321C3.45939 12.2321 2.5 13.1915 2.5 14.375V15.3571C2.5 16.5406 3.45939 17.5 4.64286 17.5H6.78571C7.96918 17.5 8.92857 16.5406 8.92857 15.3571V14.375C8.92857 13.1915 7.96918 12.2321 6.78571 12.2321H4.64286Z"
-                    fill="currentColor"
-                  />
-                  <path
-                    d="M11.0714 4.64286C11.0714 3.45939 12.0308 2.5 13.2143 2.5H15.3571C16.5406 2.5 17.5 3.45939 17.5 4.64286V5.50742C17.5 6.69088 16.5406 7.65027 15.3571 7.65027H13.2143C12.0308 7.65027 11.0714 6.69088 11.0714 5.50742V4.64286Z"
-                    fill="currentColor"
-                  />
-                  <path
-                    d="M11.0714 11.5179C11.0714 10.3344 12.0308 9.375 13.2143 9.375H15.3571C16.5406 9.375 17.5 10.3344 17.5 11.5179V15.3571C17.5 16.5406 16.5406 17.5 15.3571 17.5H13.2143C12.0308 17.5 11.0714 16.5406 11.0714 15.3571V11.5179Z"
-                    fill="currentColor"
-                  />
-                </svg>
+                <DashboardSvg />
                 Dashboard
               </a>
             </Link>
@@ -62,23 +41,10 @@ const DashboardLayout: React.FC = ({ children }) => {
               <a
                 className={[
                   'dashboard-navigation__link',
-                  router.pathname === '/employees' ? 'active' : '',
+                  router.pathname.includes('/employees') ? 'active' : '',
                 ].join(' ')}
               >
-                <svg
-                  width="20"
-                  height="20"
-                  viewBox="0 0 20 20"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    fillRule="evenodd"
-                    clipRule="evenodd"
-                    d="M13.4654 6.86087C13.4654 8.81894 11.9044 10.389 9.95769 10.389C8.01093 10.389 6.44995 8.81894 6.44995 6.86087C6.44995 4.90205 8.01093 3.3335 9.95769 3.3335C11.9044 3.3335 13.4654 4.90205 13.4654 6.86087ZM9.95772 16.6668C7.0984 16.6668 4.65694 16.2135 4.65694 14.4003C4.65694 12.5863 7.08273 12.1165 9.95772 12.1165C12.817 12.1165 15.2585 12.5698 15.2585 14.3838C15.2585 16.197 12.8327 16.6668 9.95772 16.6668ZM14.9644 6.92452C14.9644 7.92269 14.6667 8.85257 14.1443 9.62559C14.0906 9.70514 14.1384 9.81247 14.2331 9.82898C14.3637 9.85149 14.4988 9.86425 14.6361 9.868C16.0053 9.90403 17.2342 9.01768 17.5737 7.68328C18.0766 5.7012 16.6 3.92175 14.7196 3.92175C14.5152 3.92175 14.3197 3.94351 14.1294 3.98254C14.1033 3.98854 14.0757 4.00055 14.0608 4.02382C14.0429 4.05233 14.0563 4.09061 14.0742 4.11538C14.639 4.91166 14.9644 5.88282 14.9644 6.92452ZM17.232 11.2606C18.152 11.4415 18.7572 11.8108 19.0079 12.3474C19.2198 12.7879 19.2198 13.299 19.0079 13.7388C18.6243 14.5711 17.3879 14.8383 16.9074 14.9073C16.8082 14.9224 16.7283 14.836 16.7388 14.7362C16.9843 12.4299 15.0315 11.3364 14.5264 11.085C14.5047 11.0738 14.5003 11.0565 14.5025 11.046C14.504 11.0385 14.513 11.0265 14.5294 11.0242C15.6225 11.004 16.7977 11.1541 17.232 11.2606ZM5.36441 9.86792C5.5017 9.86417 5.63601 9.85216 5.76734 9.8289C5.8621 9.81239 5.90986 9.70506 5.85613 9.62551C5.33381 8.85249 5.03609 7.92261 5.03609 6.92444C5.03609 5.88274 5.36142 4.91158 5.92627 4.1153C5.94418 4.09053 5.95687 4.05225 5.9397 4.02374C5.92478 4.00122 5.89643 3.98846 5.87106 3.98246C5.68004 3.94343 5.48454 3.92167 5.28009 3.92167C3.39974 3.92167 1.92308 5.70112 2.42674 7.6832C2.76625 9.0176 3.99519 9.90395 5.36441 9.86792ZM5.49761 11.0455C5.49985 11.0568 5.49537 11.0733 5.47448 11.0853C4.96858 11.3367 3.01585 12.4302 3.26134 14.7358C3.27179 14.8364 3.1927 14.9219 3.09346 14.9076C2.61292 14.8386 1.37652 14.5714 0.99299 13.7391C0.780332 13.2986 0.780332 12.7882 0.99299 12.3477C1.2437 11.8111 1.8481 11.4418 2.76813 11.2602C3.20314 11.1544 4.37761 11.0043 5.4715 11.0245C5.48791 11.0268 5.49612 11.0388 5.49761 11.0455Z"
-                    fill="currentColor"
-                  />
-                </svg>
+                <EmployeeSvg />
                 Employee Settings
               </a>
             </Link>
@@ -91,20 +57,7 @@ const DashboardLayout: React.FC = ({ children }) => {
                   router.pathname === '/payroll' ? 'active' : '',
                 ].join(' ')}
               >
-                <svg
-                  width="20"
-                  height="20"
-                  viewBox="0 0 20 20"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    fillRule="evenodd"
-                    clipRule="evenodd"
-                    d="M7.43797 13.6582H11.9264C12.2647 13.6582 12.5452 13.3748 12.5452 13.0332C12.5452 12.6915 12.2647 12.4165 11.9264 12.4165H7.43797C7.09968 12.4165 6.81915 12.6915 6.81915 13.0332C6.81915 13.3748 7.09968 13.6582 7.43797 13.6582ZM10.2267 8.24981H7.43797C7.09968 8.24981 6.81915 8.53314 6.81915 8.87481C6.81915 9.21648 7.09968 9.49148 7.43797 9.49148H10.2267C10.565 9.49148 10.8456 9.21648 10.8456 8.87481C10.8456 8.53314 10.565 8.24981 10.2267 8.24981ZM16.1149 7.52118C16.3088 7.51894 16.52 7.5165 16.7119 7.5165C16.9182 7.5165 17.0832 7.68317 17.0832 7.8915V14.5915C17.0832 16.6582 15.4248 18.3332 13.3786 18.3332H6.81089C4.66568 18.3332 2.9165 16.5748 2.9165 14.4082V5.42484C2.9165 3.35817 4.58317 1.6665 6.63763 1.6665H11.0436C11.2581 1.6665 11.4231 1.8415 11.4231 2.04984V4.73317C11.4231 6.25817 12.669 7.50817 14.1789 7.5165C14.5316 7.5165 14.8425 7.51914 15.1146 7.52145C15.3263 7.52324 15.5145 7.52484 15.6805 7.52484C15.798 7.52484 15.9502 7.52308 16.1149 7.52118ZM16.3424 6.30485C15.6642 6.30735 14.8647 6.30485 14.2896 6.29901C13.3771 6.29901 12.6254 5.53985 12.6254 4.61818V2.42151C12.6254 2.06235 13.0569 1.88401 13.3036 2.14318C13.7506 2.6126 14.3651 3.25804 14.9765 3.90035C15.5851 4.53968 16.1908 5.17591 16.6254 5.63235C16.8663 5.88485 16.6898 6.30401 16.3424 6.30485Z"
-                    fill="currentColor"
-                  />
-                </svg>
+                <PayrollSvg />
                 Payroll
               </a>
             </Link>
@@ -117,20 +70,7 @@ const DashboardLayout: React.FC = ({ children }) => {
                   router.pathname === '/wallet-billings' ? 'active' : '',
                 ].join(' ')}
               >
-                <svg
-                  width="20"
-                  height="20"
-                  viewBox="0 0 20 20"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    fillRule="evenodd"
-                    clipRule="evenodd"
-                    d="M7.43797 13.6582H11.9264C12.2647 13.6582 12.5452 13.3748 12.5452 13.0332C12.5452 12.6915 12.2647 12.4165 11.9264 12.4165H7.43797C7.09968 12.4165 6.81915 12.6915 6.81915 13.0332C6.81915 13.3748 7.09968 13.6582 7.43797 13.6582ZM10.2267 8.24981H7.43797C7.09968 8.24981 6.81915 8.53314 6.81915 8.87481C6.81915 9.21648 7.09968 9.49148 7.43797 9.49148H10.2267C10.565 9.49148 10.8456 9.21648 10.8456 8.87481C10.8456 8.53314 10.565 8.24981 10.2267 8.24981ZM16.1149 7.52118C16.3088 7.51894 16.52 7.5165 16.7119 7.5165C16.9182 7.5165 17.0832 7.68317 17.0832 7.8915V14.5915C17.0832 16.6582 15.4248 18.3332 13.3786 18.3332H6.81089C4.66568 18.3332 2.9165 16.5748 2.9165 14.4082V5.42484C2.9165 3.35817 4.58317 1.6665 6.63763 1.6665H11.0436C11.2581 1.6665 11.4231 1.8415 11.4231 2.04984V4.73317C11.4231 6.25817 12.669 7.50817 14.1789 7.5165C14.5316 7.5165 14.8425 7.51914 15.1146 7.52145C15.3263 7.52324 15.5145 7.52484 15.6805 7.52484C15.798 7.52484 15.9502 7.52308 16.1149 7.52118ZM16.3424 6.30485C15.6642 6.30735 14.8647 6.30485 14.2896 6.29901C13.3771 6.29901 12.6254 5.53985 12.6254 4.61818V2.42151C12.6254 2.06235 13.0569 1.88401 13.3036 2.14318C13.7506 2.6126 14.3651 3.25804 14.9765 3.90035C15.5851 4.53968 16.1908 5.17591 16.6254 5.63235C16.8663 5.88485 16.6898 6.30401 16.3424 6.30485Z"
-                    fill="currentColor"
-                  />
-                </svg>
+                <WalletBillingsSvg />
                 Wallet &amp; Billings
               </a>
             </Link>
@@ -144,20 +84,7 @@ const DashboardLayout: React.FC = ({ children }) => {
                   router.pathname === '/admin-management' ? 'active' : '',
                 ].join(' ')}
               >
-                <svg
-                  width="20"
-                  height="20"
-                  viewBox="0 0 20 20"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    fillRule="evenodd"
-                    clipRule="evenodd"
-                    d="M7.91647 10.4616C10.2119 10.4616 12.052 8.59771 12.052 6.27253C12.052 3.94735 10.2119 2.0835 7.91647 2.0835C5.62099 2.0835 3.78094 3.94735 3.78094 6.27253C3.78094 8.59771 5.62099 10.4616 7.91647 10.4616ZM7.9165 12.5128C4.54502 12.5128 1.6665 13.0519 1.6665 15.2055C1.6665 17.3583 4.52751 17.9168 7.9165 17.9168C11.2871 17.9168 14.1665 17.3778 14.1665 15.2241C14.1665 13.0705 11.3055 12.5128 7.9165 12.5128ZM16.5815 7.99005H17.584C17.9967 7.99005 18.3332 8.33125 18.3332 8.74973C18.3332 9.1682 17.9967 9.50941 17.584 9.50941H16.5815V10.4905C16.5815 10.909 16.2458 11.2502 15.8323 11.2502C15.4196 11.2502 15.0832 10.909 15.0832 10.4905V9.50941H14.0823C13.6688 9.50941 13.3332 9.1682 13.3332 8.74973C13.3332 8.33125 13.6688 7.99005 14.0823 7.99005H15.0832V7.00984C15.0832 6.59137 15.4196 6.25016 15.8323 6.25016C16.2458 6.25016 16.5815 6.59137 16.5815 7.00984V7.99005Z"
-                    fill="currentColor"
-                  />
-                </svg>
+                <AdminManagementSvg />
                 Admin Management
               </a>
             </Link>
@@ -171,20 +98,7 @@ const DashboardLayout: React.FC = ({ children }) => {
                   router.pathname === '/organisation-setting' ? 'active' : '',
                 ].join(' ')}
               >
-                <svg
-                  width="20"
-                  height="20"
-                  viewBox="0 0 20 20"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    fillRule="evenodd"
-                    clipRule="evenodd"
-                    d="M17.0016 11.3165C17.2997 11.4748 17.5296 11.7248 17.6915 11.9748C18.0066 12.4915 17.981 13.1248 17.6744 13.6832L17.0783 14.6832C16.7632 15.2165 16.1755 15.5498 15.5709 15.5498C15.2728 15.5498 14.9407 15.4665 14.6682 15.2998C14.4467 15.1582 14.1912 15.1082 13.9187 15.1082C13.0756 15.1082 12.3688 15.7998 12.3432 16.6248C12.3432 17.5832 11.5597 18.3332 10.5803 18.3332H9.42211C8.43422 18.3332 7.65072 17.5832 7.65072 16.6248C7.63369 15.7998 6.92683 15.1082 6.08372 15.1082C5.80268 15.1082 5.54719 15.1582 5.33428 15.2998C5.06176 15.4665 4.72111 15.5498 4.43155 15.5498C3.81838 15.5498 3.23075 15.2165 2.91565 14.6832L2.32802 13.6832C2.01292 13.1415 1.99589 12.4915 2.31099 11.9748C2.44725 11.7248 2.70274 11.4748 2.9923 11.3165C3.23075 11.1998 3.38405 11.0082 3.52882 10.7832C3.95464 10.0665 3.69915 9.12484 2.97526 8.69984C2.13215 8.22484 1.85963 7.1665 2.34506 6.3415L2.91565 5.35817C3.4096 4.53317 4.46562 4.2415 5.31725 4.72484C6.05817 5.12484 7.02051 4.85817 7.45484 4.14984C7.5911 3.9165 7.66775 3.6665 7.65072 3.4165C7.63369 3.0915 7.72737 2.78317 7.88918 2.53317C8.20428 2.0165 8.77487 1.68317 9.39656 1.6665H10.5974C11.2276 1.6665 11.7982 2.0165 12.1133 2.53317C12.2666 2.78317 12.3688 3.0915 12.3432 3.4165C12.3262 3.6665 12.4028 3.9165 12.5391 4.14984C12.9734 4.85817 13.9358 5.12484 14.6852 4.72484C15.5283 4.2415 16.5928 4.53317 17.0783 5.35817L17.6489 6.3415C18.1428 7.1665 17.8703 8.22484 17.0187 8.69984C16.2948 9.12484 16.0393 10.0665 16.4736 10.7832C16.6099 11.0082 16.7632 11.1998 17.0016 11.3165ZM7.59108 10.0082C7.59108 11.3165 8.67265 12.3582 10.0097 12.3582C11.3468 12.3582 12.4028 11.3165 12.4028 10.0082C12.4028 8.69984 11.3468 7.64984 10.0097 7.64984C8.67265 7.64984 7.59108 8.69984 7.59108 10.0082Z"
-                    fill="currentColor"
-                  />
-                </svg>
+                <OrganizationSettingsSvg />
                 Organisation Setting
               </a>
             </Link>
@@ -198,20 +112,7 @@ const DashboardLayout: React.FC = ({ children }) => {
                   router.pathname === '/remittances' ? 'active' : '',
                 ].join(' ')}
               >
-                <svg
-                  width="20"
-                  height="20"
-                  viewBox="0 0 20 20"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    fillRule="evenodd"
-                    clipRule="evenodd"
-                    d="M16.4673 10.0008C16.4673 10.6787 17.024 11.2303 17.7082 11.2303C18.0532 11.2303 18.3332 11.5078 18.3332 11.8496V14.0799C18.3332 15.9658 16.7848 17.5 14.8815 17.5H5.119C3.21567 17.5 1.6665 15.9658 1.6665 14.0799V11.8496C1.6665 11.5078 1.9465 11.2303 2.2915 11.2303C2.9765 11.2303 3.53317 10.6787 3.53317 10.0008C3.53317 9.34025 2.999 8.84317 2.2915 8.84317C2.12567 8.84317 1.96734 8.77794 1.84984 8.66151C1.73234 8.54508 1.6665 8.38737 1.6665 8.22388L1.66817 5.92095C1.66817 4.03501 3.2165 2.5 5.11984 2.5H14.8798C16.7832 2.5 18.3323 4.03501 18.3323 5.92095L18.3332 8.15204C18.3332 8.31553 18.2673 8.47407 18.1507 8.58967C18.0332 8.7061 17.8748 8.77133 17.7082 8.77133C17.024 8.77133 16.4673 9.32291 16.4673 10.0008ZM11.8765 10.54L12.859 9.59208C13.0298 9.42859 13.089 9.18748 13.0148 8.96453C12.9415 8.74159 12.7498 8.58305 12.5182 8.55085L11.1607 8.35433L10.5532 7.13557C10.449 6.92583 10.2373 6.79537 10.0015 6.79455H9.99985C9.76485 6.79455 9.55318 6.92501 9.44735 7.13474L8.83985 8.35433L7.48485 8.55002C7.25068 8.58305 7.05901 8.74159 6.98485 8.96453C6.91151 9.18748 6.97068 9.42859 7.14068 9.59208L8.12318 10.54L7.89151 11.8801C7.85151 12.1113 7.94568 12.3409 8.13735 12.4788C8.24568 12.5556 8.37151 12.5952 8.49901 12.5952C8.59651 12.5952 8.69485 12.5713 8.78485 12.5242L9.99985 11.8917L11.2123 12.5226C11.4223 12.634 11.6715 12.6167 11.8623 12.478C12.0548 12.3409 12.149 12.1113 12.109 11.8801L11.8765 10.54Z"
-                    fill="currentColor"
-                  />
-                </svg>
+                <RemittancesSvg />
                 Remittances
               </a>
             </Link>
@@ -225,20 +126,7 @@ const DashboardLayout: React.FC = ({ children }) => {
                   router.pathname === '/audit-trail' ? 'active' : '',
                 ].join(' ')}
               >
-                <svg
-                  width="20"
-                  height="20"
-                  viewBox="0 0 20 20"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    fillRule="evenodd"
-                    clipRule="evenodd"
-                    d="M9.77325 18.2612C9.86553 18.3094 9.96872 18.334 10.0719 18.3331C10.1751 18.3323 10.2775 18.307 10.3706 18.258L13.3438 16.6686C14.187 16.2191 14.8472 15.7166 15.3623 15.1315C16.4823 13.8567 17.0939 12.2297 17.083 10.552L17.0478 5.01816C17.0444 4.38076 16.6258 3.81201 16.0066 3.60527L10.4755 1.74947C10.1424 1.6367 9.77744 1.63916 9.45025 1.75519L3.94004 3.67718C3.32425 3.8921 2.91317 4.46493 2.91652 5.10314L2.95176 10.6329C2.96267 12.313 3.59523 13.9327 4.73369 15.1944C5.25383 15.7713 5.91996 16.2665 6.77233 16.7086L9.77325 18.2612ZM8.98613 11.7573C9.1103 11.8766 9.27138 11.9354 9.43245 11.9338C9.59353 11.933 9.75377 11.8725 9.87626 11.7516L13.1255 8.54825C13.3696 8.30718 13.3671 7.91985 13.1205 7.68205C12.873 7.44425 12.4745 7.44589 12.2303 7.68695L9.42323 10.4539L8.27387 9.34908C8.02638 9.11128 7.62872 9.11373 7.38374 9.3548C7.13961 9.59586 7.14213 9.9832 7.38962 10.221L8.98613 11.7573Z"
-                    fill="currentColor"
-                  />
-                </svg>
+                <AuditTrailSvg />
                 Audit Trail
               </a>
             </Link>
@@ -262,3 +150,149 @@ const DashboardLayout: React.FC = ({ children }) => {
 };
 
 export default DashboardLayout;
+
+const DashboardSvg = () => (
+  <svg
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M4.64286 2.5C3.45939 2.5 2.5 3.45939 2.5 4.64286V8.48214C2.5 9.66561 3.45939 10.625 4.64286 10.625H6.78571C7.96918 10.625 8.92857 9.66561 8.92857 8.48214V4.64286C8.92857 3.45939 7.96918 2.5 6.78571 2.5H4.64286Z"
+      fill="currentColor"
+    />
+    <path
+      d="M4.64286 12.2321C3.45939 12.2321 2.5 13.1915 2.5 14.375V15.3571C2.5 16.5406 3.45939 17.5 4.64286 17.5H6.78571C7.96918 17.5 8.92857 16.5406 8.92857 15.3571V14.375C8.92857 13.1915 7.96918 12.2321 6.78571 12.2321H4.64286Z"
+      fill="currentColor"
+    />
+    <path
+      d="M11.0714 4.64286C11.0714 3.45939 12.0308 2.5 13.2143 2.5H15.3571C16.5406 2.5 17.5 3.45939 17.5 4.64286V5.50742C17.5 6.69088 16.5406 7.65027 15.3571 7.65027H13.2143C12.0308 7.65027 11.0714 6.69088 11.0714 5.50742V4.64286Z"
+      fill="currentColor"
+    />
+    <path
+      d="M11.0714 11.5179C11.0714 10.3344 12.0308 9.375 13.2143 9.375H15.3571C16.5406 9.375 17.5 10.3344 17.5 11.5179V15.3571C17.5 16.5406 16.5406 17.5 15.3571 17.5H13.2143C12.0308 17.5 11.0714 16.5406 11.0714 15.3571V11.5179Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+
+const EmployeeSvg = () => (
+  <svg
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M13.4654 6.86087C13.4654 8.81894 11.9044 10.389 9.95769 10.389C8.01093 10.389 6.44995 8.81894 6.44995 6.86087C6.44995 4.90205 8.01093 3.3335 9.95769 3.3335C11.9044 3.3335 13.4654 4.90205 13.4654 6.86087ZM9.95772 16.6668C7.0984 16.6668 4.65694 16.2135 4.65694 14.4003C4.65694 12.5863 7.08273 12.1165 9.95772 12.1165C12.817 12.1165 15.2585 12.5698 15.2585 14.3838C15.2585 16.197 12.8327 16.6668 9.95772 16.6668ZM14.9644 6.92452C14.9644 7.92269 14.6667 8.85257 14.1443 9.62559C14.0906 9.70514 14.1384 9.81247 14.2331 9.82898C14.3637 9.85149 14.4988 9.86425 14.6361 9.868C16.0053 9.90403 17.2342 9.01768 17.5737 7.68328C18.0766 5.7012 16.6 3.92175 14.7196 3.92175C14.5152 3.92175 14.3197 3.94351 14.1294 3.98254C14.1033 3.98854 14.0757 4.00055 14.0608 4.02382C14.0429 4.05233 14.0563 4.09061 14.0742 4.11538C14.639 4.91166 14.9644 5.88282 14.9644 6.92452ZM17.232 11.2606C18.152 11.4415 18.7572 11.8108 19.0079 12.3474C19.2198 12.7879 19.2198 13.299 19.0079 13.7388C18.6243 14.5711 17.3879 14.8383 16.9074 14.9073C16.8082 14.9224 16.7283 14.836 16.7388 14.7362C16.9843 12.4299 15.0315 11.3364 14.5264 11.085C14.5047 11.0738 14.5003 11.0565 14.5025 11.046C14.504 11.0385 14.513 11.0265 14.5294 11.0242C15.6225 11.004 16.7977 11.1541 17.232 11.2606ZM5.36441 9.86792C5.5017 9.86417 5.63601 9.85216 5.76734 9.8289C5.8621 9.81239 5.90986 9.70506 5.85613 9.62551C5.33381 8.85249 5.03609 7.92261 5.03609 6.92444C5.03609 5.88274 5.36142 4.91158 5.92627 4.1153C5.94418 4.09053 5.95687 4.05225 5.9397 4.02374C5.92478 4.00122 5.89643 3.98846 5.87106 3.98246C5.68004 3.94343 5.48454 3.92167 5.28009 3.92167C3.39974 3.92167 1.92308 5.70112 2.42674 7.6832C2.76625 9.0176 3.99519 9.90395 5.36441 9.86792ZM5.49761 11.0455C5.49985 11.0568 5.49537 11.0733 5.47448 11.0853C4.96858 11.3367 3.01585 12.4302 3.26134 14.7358C3.27179 14.8364 3.1927 14.9219 3.09346 14.9076C2.61292 14.8386 1.37652 14.5714 0.99299 13.7391C0.780332 13.2986 0.780332 12.7882 0.99299 12.3477C1.2437 11.8111 1.8481 11.4418 2.76813 11.2602C3.20314 11.1544 4.37761 11.0043 5.4715 11.0245C5.48791 11.0268 5.49612 11.0388 5.49761 11.0455Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+
+const PayrollSvg = () => (
+  <svg
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M7.43797 13.6582H11.9264C12.2647 13.6582 12.5452 13.3748 12.5452 13.0332C12.5452 12.6915 12.2647 12.4165 11.9264 12.4165H7.43797C7.09968 12.4165 6.81915 12.6915 6.81915 13.0332C6.81915 13.3748 7.09968 13.6582 7.43797 13.6582ZM10.2267 8.24981H7.43797C7.09968 8.24981 6.81915 8.53314 6.81915 8.87481C6.81915 9.21648 7.09968 9.49148 7.43797 9.49148H10.2267C10.565 9.49148 10.8456 9.21648 10.8456 8.87481C10.8456 8.53314 10.565 8.24981 10.2267 8.24981ZM16.1149 7.52118C16.3088 7.51894 16.52 7.5165 16.7119 7.5165C16.9182 7.5165 17.0832 7.68317 17.0832 7.8915V14.5915C17.0832 16.6582 15.4248 18.3332 13.3786 18.3332H6.81089C4.66568 18.3332 2.9165 16.5748 2.9165 14.4082V5.42484C2.9165 3.35817 4.58317 1.6665 6.63763 1.6665H11.0436C11.2581 1.6665 11.4231 1.8415 11.4231 2.04984V4.73317C11.4231 6.25817 12.669 7.50817 14.1789 7.5165C14.5316 7.5165 14.8425 7.51914 15.1146 7.52145C15.3263 7.52324 15.5145 7.52484 15.6805 7.52484C15.798 7.52484 15.9502 7.52308 16.1149 7.52118ZM16.3424 6.30485C15.6642 6.30735 14.8647 6.30485 14.2896 6.29901C13.3771 6.29901 12.6254 5.53985 12.6254 4.61818V2.42151C12.6254 2.06235 13.0569 1.88401 13.3036 2.14318C13.7506 2.6126 14.3651 3.25804 14.9765 3.90035C15.5851 4.53968 16.1908 5.17591 16.6254 5.63235C16.8663 5.88485 16.6898 6.30401 16.3424 6.30485Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+
+const WalletBillingsSvg = () => (
+  <svg
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M7.43797 13.6582H11.9264C12.2647 13.6582 12.5452 13.3748 12.5452 13.0332C12.5452 12.6915 12.2647 12.4165 11.9264 12.4165H7.43797C7.09968 12.4165 6.81915 12.6915 6.81915 13.0332C6.81915 13.3748 7.09968 13.6582 7.43797 13.6582ZM10.2267 8.24981H7.43797C7.09968 8.24981 6.81915 8.53314 6.81915 8.87481C6.81915 9.21648 7.09968 9.49148 7.43797 9.49148H10.2267C10.565 9.49148 10.8456 9.21648 10.8456 8.87481C10.8456 8.53314 10.565 8.24981 10.2267 8.24981ZM16.1149 7.52118C16.3088 7.51894 16.52 7.5165 16.7119 7.5165C16.9182 7.5165 17.0832 7.68317 17.0832 7.8915V14.5915C17.0832 16.6582 15.4248 18.3332 13.3786 18.3332H6.81089C4.66568 18.3332 2.9165 16.5748 2.9165 14.4082V5.42484C2.9165 3.35817 4.58317 1.6665 6.63763 1.6665H11.0436C11.2581 1.6665 11.4231 1.8415 11.4231 2.04984V4.73317C11.4231 6.25817 12.669 7.50817 14.1789 7.5165C14.5316 7.5165 14.8425 7.51914 15.1146 7.52145C15.3263 7.52324 15.5145 7.52484 15.6805 7.52484C15.798 7.52484 15.9502 7.52308 16.1149 7.52118ZM16.3424 6.30485C15.6642 6.30735 14.8647 6.30485 14.2896 6.29901C13.3771 6.29901 12.6254 5.53985 12.6254 4.61818V2.42151C12.6254 2.06235 13.0569 1.88401 13.3036 2.14318C13.7506 2.6126 14.3651 3.25804 14.9765 3.90035C15.5851 4.53968 16.1908 5.17591 16.6254 5.63235C16.8663 5.88485 16.6898 6.30401 16.3424 6.30485Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+
+const AdminManagementSvg = () => (
+  <svg
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M7.91647 10.4616C10.2119 10.4616 12.052 8.59771 12.052 6.27253C12.052 3.94735 10.2119 2.0835 7.91647 2.0835C5.62099 2.0835 3.78094 3.94735 3.78094 6.27253C3.78094 8.59771 5.62099 10.4616 7.91647 10.4616ZM7.9165 12.5128C4.54502 12.5128 1.6665 13.0519 1.6665 15.2055C1.6665 17.3583 4.52751 17.9168 7.9165 17.9168C11.2871 17.9168 14.1665 17.3778 14.1665 15.2241C14.1665 13.0705 11.3055 12.5128 7.9165 12.5128ZM16.5815 7.99005H17.584C17.9967 7.99005 18.3332 8.33125 18.3332 8.74973C18.3332 9.1682 17.9967 9.50941 17.584 9.50941H16.5815V10.4905C16.5815 10.909 16.2458 11.2502 15.8323 11.2502C15.4196 11.2502 15.0832 10.909 15.0832 10.4905V9.50941H14.0823C13.6688 9.50941 13.3332 9.1682 13.3332 8.74973C13.3332 8.33125 13.6688 7.99005 14.0823 7.99005H15.0832V7.00984C15.0832 6.59137 15.4196 6.25016 15.8323 6.25016C16.2458 6.25016 16.5815 6.59137 16.5815 7.00984V7.99005Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+
+const OrganizationSettingsSvg = () => (
+  <svg
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M17.0016 11.3165C17.2997 11.4748 17.5296 11.7248 17.6915 11.9748C18.0066 12.4915 17.981 13.1248 17.6744 13.6832L17.0783 14.6832C16.7632 15.2165 16.1755 15.5498 15.5709 15.5498C15.2728 15.5498 14.9407 15.4665 14.6682 15.2998C14.4467 15.1582 14.1912 15.1082 13.9187 15.1082C13.0756 15.1082 12.3688 15.7998 12.3432 16.6248C12.3432 17.5832 11.5597 18.3332 10.5803 18.3332H9.42211C8.43422 18.3332 7.65072 17.5832 7.65072 16.6248C7.63369 15.7998 6.92683 15.1082 6.08372 15.1082C5.80268 15.1082 5.54719 15.1582 5.33428 15.2998C5.06176 15.4665 4.72111 15.5498 4.43155 15.5498C3.81838 15.5498 3.23075 15.2165 2.91565 14.6832L2.32802 13.6832C2.01292 13.1415 1.99589 12.4915 2.31099 11.9748C2.44725 11.7248 2.70274 11.4748 2.9923 11.3165C3.23075 11.1998 3.38405 11.0082 3.52882 10.7832C3.95464 10.0665 3.69915 9.12484 2.97526 8.69984C2.13215 8.22484 1.85963 7.1665 2.34506 6.3415L2.91565 5.35817C3.4096 4.53317 4.46562 4.2415 5.31725 4.72484C6.05817 5.12484 7.02051 4.85817 7.45484 4.14984C7.5911 3.9165 7.66775 3.6665 7.65072 3.4165C7.63369 3.0915 7.72737 2.78317 7.88918 2.53317C8.20428 2.0165 8.77487 1.68317 9.39656 1.6665H10.5974C11.2276 1.6665 11.7982 2.0165 12.1133 2.53317C12.2666 2.78317 12.3688 3.0915 12.3432 3.4165C12.3262 3.6665 12.4028 3.9165 12.5391 4.14984C12.9734 4.85817 13.9358 5.12484 14.6852 4.72484C15.5283 4.2415 16.5928 4.53317 17.0783 5.35817L17.6489 6.3415C18.1428 7.1665 17.8703 8.22484 17.0187 8.69984C16.2948 9.12484 16.0393 10.0665 16.4736 10.7832C16.6099 11.0082 16.7632 11.1998 17.0016 11.3165ZM7.59108 10.0082C7.59108 11.3165 8.67265 12.3582 10.0097 12.3582C11.3468 12.3582 12.4028 11.3165 12.4028 10.0082C12.4028 8.69984 11.3468 7.64984 10.0097 7.64984C8.67265 7.64984 7.59108 8.69984 7.59108 10.0082Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+
+const RemittancesSvg = () => (
+  <svg
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M16.4673 10.0008C16.4673 10.6787 17.024 11.2303 17.7082 11.2303C18.0532 11.2303 18.3332 11.5078 18.3332 11.8496V14.0799C18.3332 15.9658 16.7848 17.5 14.8815 17.5H5.119C3.21567 17.5 1.6665 15.9658 1.6665 14.0799V11.8496C1.6665 11.5078 1.9465 11.2303 2.2915 11.2303C2.9765 11.2303 3.53317 10.6787 3.53317 10.0008C3.53317 9.34025 2.999 8.84317 2.2915 8.84317C2.12567 8.84317 1.96734 8.77794 1.84984 8.66151C1.73234 8.54508 1.6665 8.38737 1.6665 8.22388L1.66817 5.92095C1.66817 4.03501 3.2165 2.5 5.11984 2.5H14.8798C16.7832 2.5 18.3323 4.03501 18.3323 5.92095L18.3332 8.15204C18.3332 8.31553 18.2673 8.47407 18.1507 8.58967C18.0332 8.7061 17.8748 8.77133 17.7082 8.77133C17.024 8.77133 16.4673 9.32291 16.4673 10.0008ZM11.8765 10.54L12.859 9.59208C13.0298 9.42859 13.089 9.18748 13.0148 8.96453C12.9415 8.74159 12.7498 8.58305 12.5182 8.55085L11.1607 8.35433L10.5532 7.13557C10.449 6.92583 10.2373 6.79537 10.0015 6.79455H9.99985C9.76485 6.79455 9.55318 6.92501 9.44735 7.13474L8.83985 8.35433L7.48485 8.55002C7.25068 8.58305 7.05901 8.74159 6.98485 8.96453C6.91151 9.18748 6.97068 9.42859 7.14068 9.59208L8.12318 10.54L7.89151 11.8801C7.85151 12.1113 7.94568 12.3409 8.13735 12.4788C8.24568 12.5556 8.37151 12.5952 8.49901 12.5952C8.59651 12.5952 8.69485 12.5713 8.78485 12.5242L9.99985 11.8917L11.2123 12.5226C11.4223 12.634 11.6715 12.6167 11.8623 12.478C12.0548 12.3409 12.149 12.1113 12.109 11.8801L11.8765 10.54Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+
+const AuditTrailSvg = () => (
+  <svg
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M9.77325 18.2612C9.86553 18.3094 9.96872 18.334 10.0719 18.3331C10.1751 18.3323 10.2775 18.307 10.3706 18.258L13.3438 16.6686C14.187 16.2191 14.8472 15.7166 15.3623 15.1315C16.4823 13.8567 17.0939 12.2297 17.083 10.552L17.0478 5.01816C17.0444 4.38076 16.6258 3.81201 16.0066 3.60527L10.4755 1.74947C10.1424 1.6367 9.77744 1.63916 9.45025 1.75519L3.94004 3.67718C3.32425 3.8921 2.91317 4.46493 2.91652 5.10314L2.95176 10.6329C2.96267 12.313 3.59523 13.9327 4.73369 15.1944C5.25383 15.7713 5.91996 16.2665 6.77233 16.7086L9.77325 18.2612ZM8.98613 11.7573C9.1103 11.8766 9.27138 11.9354 9.43245 11.9338C9.59353 11.933 9.75377 11.8725 9.87626 11.7516L13.1255 8.54825C13.3696 8.30718 13.3671 7.91985 13.1205 7.68205C12.873 7.44425 12.4745 7.44589 12.2303 7.68695L9.42323 10.4539L8.27387 9.34908C8.02638 9.11128 7.62872 9.11373 7.38374 9.3548C7.13961 9.59586 7.14213 9.9832 7.38962 10.221L8.98613 11.7573Z"
+      fill="currentColor"
+    />
+  </svg>
+);
